Move CreateThingDialog to new file
